Updated lib/.
