+++ lib/index.js
@@ -144,6 +144,7 @@ class Storage extends _nxusCore.NxusModule {
     this.waterlineConfig = null;
     this.collections = {};
     this.connections = null;
+    this._adapters = {};
 
     _nxusCore.application.once('init', () => {
       return _bluebird2.default.all([this._setupAdapter()]);
@@ -210,6 +211,7 @@ class Storage extends _nxusCore.NxusModule {
     try {
       fs.accessSync(dir);
     } catch (e) {
+      this.log.debug('modelDir skipping:', dir, e.message);
       return;
     }
     let identities = [];
@@ -221,6 +223,7 @@ class Storage extends _nxusCore.NxusModule {
           m = m.default;
         }
         identities.push(m.prototype.identity);
+        this.log.debug('modelDir adding model', file);
         return this.provide('model', m);
       }
     }).then(() => {
@@ -233,16 +236,20 @@ class Storage extends _nxusCore.NxusModule {
   _setupAdapter() {
     for (var key in this.config.adapters) {
       if (_underscore2.default.isString(this.config.adapters[key])) {
-        var adapter = require(this.config.adapters[key]);
-        adapter._name = this.config.adapters[key];
-        this.config.adapters[key] = adapter;
+        try {
+          var adapter = require(this.config.adapters[key]);
+          adapter['_name'] = this.config.adapters[key];
+          this._adapters[key] = adapter;
+        } catch (err) {
+          this.log.error('_setupAdapter config for', key, 'adapter:', this.config.adapters[key], 'not found in installed dependencies.');
+        }
       }
     }
   }
 
   _disconnectDb() {
-    var adapters = _underscore2.default.values(this.config.adapters).map(e => e['_name']);
-    return _bluebird2.default.all(_underscore2.default.values(this.config.adapters), adapter => {
+    var adapterNames = _underscore2.default.values(this._adapters).map(e => e['_name']);
+    return _bluebird2.default.all(_underscore2.default.values(this._adapters), adapter => {
       return new _bluebird2.default(resolve => {
         adapter.teardown(null, resolve);
       });
@@ -252,9 +259,9 @@ class Storage extends _nxusCore.NxusModule {
       return new _bluebird2.default(resolve => {
         // we only want to reload nxus code
         // but we need to always reload mongoose so that models can be rebuilt
-        adapters = new RegExp("^.*(" + adapters.join("|") + ").*.js");
+        let adaptersExp = new RegExp("^.*(" + adapterNames.join("|") + ")\\" + _path2.default.sep + ".*.js");
         _underscore2.default.each(require.cache, (v, k) => {
-          if (!adapters.test(k)) return;
+          if (!adaptersExp.test(k)) return;
           delete require.cache[k];
         });
         resolve();
@@ -265,7 +272,7 @@ class Storage extends _nxusCore.NxusModule {
   _connectDb() {
     this.log.debug('Connecting to dB', this.config.connections);
     return this.waterline.initializeAsync({
-      adapters: this.config.adapters,
+      adapters: this._adapters,
       connections: this.config.connections,
       defaults: this.config.defaults
     }).then(obj => {
@@ -277,7 +284,7 @@ class Storage extends _nxusCore.NxusModule {
   }
 
   emitModelEvent(action, identity, record) {
-    this.log.debug('Emitting model event', action, identity);
+    this.log.trace('Emitting model event', action, identity);
     this.emit('model.' + action, identity, record);
     this.emit('model.' + action + '.' + identity, identity, record);
   }
