+++ lib/test/index.js
@@ -21,41 +21,6 @@ var _Two2 = _interopRequireDefault(_Two);
 
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
 
-// example from RFC 7946, section 1.5 (deconstructed and reassembled)
-const geoJSONPoint = { 'type': 'Point', 'coordinates': [102.0, 0.5] },
-      geoJSONLineString = { 'type': 'LineString',
-  'coordinates': [[102.0, 0.0], [103.0, 1.0], [104.0, 0.0], [105.0, 1.0]] },
-      geoJSONPolygon = { 'type': 'Polygon',
-  'coordinates': [[[100.0, 0.0], [101.0, 0.0], [101.0, 1.0], [100.0, 1.0], [100.0, 0.0]]] },
-      geoJSON = { 'type': 'FeatureCollection',
-  'features': [{ 'type': 'Feature',
-    'geometry': geoJSONPoint,
-    'properties': { 'prop0': 'value0' } }, { 'type': 'Feature',
-    'geometry': geoJSONLineString,
-    'properties': { 'prop0': 'value0', 'prop1': 0.0 } }, { 'type': 'Feature',
-    'geometry': geoJSONPolygon,
-    'properties': { 'prop0': 'value0', 'prop1': { 'this': 'that' } } }]
-},
-      geoJSONGeometry = { 'type': 'GeometryCollection',
-  'geometries': [geoJSONPolygon, geoJSONPoint, geoJSONLineString] },
-      geoJSONCentroid = { 'type': 'Point', 'coordinates': [102.0, 0.5] };
-
-// does not follow the right-hand rule; missing closure point
-const geoJSONMess = { 'type': 'Polygon',
-  'coordinates': [[[-78.81248474121094, 35.68853320738875], [-78.81248474121094, 35.862343734896484], [-78.45817565917967, 35.862343734896484], [-78.45817565917967, 35.68853320738875]]] },
-      geoJSONCleaned = { 'type': 'Polygon',
-  'coordinates': [[[-78.81248474121094, 35.68853320738875], [-78.45817565917967, 35.68853320738875], [-78.45817565917967, 35.862343734896484], [-78.81248474121094, 35.862343734896484], [-78.81248474121094, 35.68853320738875]]] };
-
-// these are slightly larger than the area covered by the geoJSONGeometry - seems to be necessary; probably issues with spherical geometry
-const surroundCoord = { 'type': 'Polygon',
-  'coordinates': [[[99.5, -0.5], [105.5, -0.5], [105.5, 1.5], [99.5, 1.5], [99.5, -0.5]]] };
-const intersectCoord = { 'type': 'Polygon',
-  'coordinates': [[[100.0, 0.0], [101.0, 0.0], [101.0, 1.0], [100.0, 1.0], [100.0, 0.0]]] };
-const throughCoord = { 'type': 'LineString',
-  'coordinates': [[100.0, 0.5], [105.0, 0.5]] };
-const multiSurroundCoord = { 'type': 'MultiPolygon',
-  'coordinates': [[[[99.5, -0.5], [101.5, -0.5], [101.5, 1.5], [99.5, 1.5], [99.5, -0.5]]], [[[101.5, -0.5], [105.5, -0.5], [105.5, 1.5], [101.5, 1.5], [101.5, -0.5]]]] };
-
 describe("Storage", () => {
   var storage;
 
@@ -66,20 +31,7 @@ describe("Storage", () => {
     _sinon2.default.spy(_.storage, "request");
   });
 
-  // beforeEach(() => {
-  //   app = new TestApp();
-  //   app.config.storage = {
-  //     adapters: {
-  //       "default": "sails-memory"
-  //     },
-  //     modelsDir: './src/models',
-  //     connections: {
-  //       'default': {
-  //         adapter: 'default'
-  //       }
-  //     }
-  //   }
-  // });
+  //nb: these tests are order dependent right now
 
   describe("Load", () => {
     it("should not be null", () => _2.default.should.not.be.null);
@@ -92,25 +44,28 @@ describe("Storage", () => {
     });
 
     it("should be instantiated", () => {
-      storage = new _2.default(_nxusCore.application);
+      storage = new _2.default();
       storage.should.not.be.null;
     });
   });
 
   describe("Init", () => {
-    it("should register for app lifecycle", () => {
-      _nxusCore.application.once.called.should.be.true;
-      _nxusCore.application.onceAfter.calledWith('load').should.be.true;
-      _nxusCore.application.once.calledWith('init').should.be.true;
-      _nxusCore.application.once.calledWith('stop').should.be.true;
-    });
-
-    it("should have config after load", () => {
-      return _nxusCore.application.emit('load').then(() => {
+    it("should register for app lifecycle", () => _nxusCore.application.once.called.should.be.true);
+    it("should register app init event", () => _nxusCore.application.once.calledWith('init').should.be.true);
+    it("should register app load event", () => _nxusCore.application.onceAfter.calledWith('load').should.be.true);
+    it("should register app stop event", () => _nxusCore.application.once.calledWith('stop').should.be.true);
+    //fire both the init and load, then check state of internal configuration
+    it("should have config after application 'init' and 'load'", () => {
+      return _nxusCore.application.emit('init').then(() => {
+        return _nxusCore.application.emit('load');
+      }).then(() => {
+        storage._adapters.should.have.property('default');
         storage.should.have.property('config');
         storage.config.should.have.property('connections');
+        storage.should.have.property('connections');
+        storage.connections.should.not.be.null;
       });
-    });
+    }).timeout(3000);
   });
   describe("Models", () => {
     before(() => {
@@ -136,6 +91,7 @@ describe("Storage", () => {
           name: 'string'
         }
       });
+
       // Shortcut around gather stub
       storage.model(Dummy);
       storage._setupAdapter();
@@ -209,7 +165,7 @@ describe("Storage", () => {
     });
 
     it("should have required the adapter", done => {
-      storage.config.adapters["default"].should.have.property("identity", "sails-memory");
+      storage._adapters["default"].should.have.property("identity", "sails-memory");
       done();
     });
   });
@@ -269,262 +225,6 @@ describe("Storage", () => {
       }).then(() => {
         storage.emit.calledWith('model.destroy').should.be.true;
         storage.emit.calledWith('model.destroy.one').should.be.true;
-      });
-    });
-  });
-  describe("Model Geo class", () => {
-    const Geo = _.GeoModel.extend({
-      identity: 'geo',
-      attributes: {
-        'name': 'string',
-        'location': 'json',
-        'locationFeatures': 'json'
-      },
-      geometryField: 'location',
-      geometryFeatureField: 'locationFeatures'
-    });
-    const record = {
-      'location': geoJSON };
-
-    beforeEach(() => {
-      storage.config = {
-        adapters: {
-          "default": "sails-mongo"
-        },
-        connections: {
-          'default': {
-            adapter: 'default',
-            url: 'mongodb://localhost/nxus-storage-test'
-          }
-        }
-      };
-      storage._setupAdapter();
-      storage.model(Geo);
-      return storage._connectDb().then(() => {
-        var geo = storage.getModel('geo');
-        return geo.destroy({});
-      });
-    });
-
-    afterEach(() => {
-      return storage._disconnectDb();
-    });
-
-    it("should return models with correct methods inherited", () => {
-      var geo = storage.getModel('geo');
-      geo.should.have.property('createGeoIndex');
-      geo.should.have.property('findWithin');
-      geo.should.have.property('findIntersects');
-    });
-
-    it("should create Geo index", () => {
-      var geo = storage.getModel('geo');
-      return geo.createGeoIndex();
-    });
-
-    it("should emit CRUD events", () => {
-      var geo = storage.getModel('geo');
-      return geo.create(record).then(obj => {
-        storage.emit.calledWith('model.create').should.be.true;
-        storage.emit.calledWith('model.create.geo').should.be.true;
-        obj.location = null;
-        return obj.save().then(() => obj); // save doesn't return object as of waterline 0.11.0
-      }).then(obj => {
-        storage.emit.calledWith('model.update').should.be.true;
-        storage.emit.calledWith('model.update.geo').should.be.true;
-        return obj.destroy();
-      }).then(() => {
-        storage.emit.calledWith('model.destroy').should.be.true;
-        storage.emit.calledWith('model.destroy.geo').should.be.true;
-      });
-    });
-
-    it("should manage features property", () => {
-      var geo = storage.getModel('geo');
-      return geo.create(record).then(obj => {
-        expect(obj).to.exist;
-        obj.should.have.property('id');
-        obj.should.have.property('createdAt');
-        obj.should.have.property('updatedAt');
-        expect(obj.location).to.deep.equal(geoJSON);
-        expect(obj.locationFeatures).to.deep.equal(geoJSONGeometry);
-        obj.location = null;
-        return obj.save().then(() => geo.findOne({ id: obj.id }));
-      }).then(obj => {
-        expect(obj).to.exist;
-        expect(obj.location).to.be.null;
-        expect(obj.locationFeatures).to.be.null;
-        return obj.destroy().then(() => geo.findOne({ id: obj.id }));
-      }).then(obj => {
-        expect(obj).to.be.undefined;
-      });
-    });
-
-    describe("Model Geo class findWithin() and findIntersects() methods", () => {
-      var geo, obj;
-
-      beforeEach(() => {
-        geo = storage.getModel('geo');
-        return geo.create(record).then(rslt => {
-          obj = rslt;
-        });
-      });
-      afterEach(() => {
-        return obj.destroy();
-      });
-
-      it("findWithin() should find entity within coordinates", () => {
-        return geo.findWithin(surroundCoord).then(query => {
-          return query();
-        }).then(rslts => {
-          expect(rslts).to.be.instanceof(Array);
-          rslts.should.have.property('length', 1);
-        });
-      });
-
-      it("findWithin() should find entity within MultiPolygon coordinates", () => {
-        return geo.findWithin(multiSurroundCoord).then(query => {
-          return query();
-        }).then(rslts => {
-          expect(rslts).to.be.instanceof(Array);
-          rslts.should.have.property('length', 1);
-        });
-      });
-
-      it("findWithin() should not find entity intersecting coordinates", () => {
-        return geo.findWithin(intersectCoord).then(query => {
-          return query();
-        }).then(rslts => {
-          expect(rslts).to.be.instanceof(Array);
-          rslts.should.have.property('length', 0);
-        });
-      });
-
-      it("findIntersects() should find entity within coordinates", () => {
-        return geo.findIntersects(surroundCoord).then(query => {
-          return query();
-        }).then(rslts => {
-          expect(rslts).to.be.instanceof(Array);
-          rslts.should.have.property('length', 1);
-        });
-      });
-
-      it("findIntersects() should find entity intersecting coordinates", () => {
-        return geo.findIntersects(intersectCoord).then(query => {
-          return query();
-        }).then(rslts => {
-          expect(rslts).to.be.instanceof(Array);
-          rslts.should.have.property('length', 1);
-        });
-      });
-
-      it("findIntersects() should find entity lined through coordinates", () => {
-        return geo.findIntersects(throughCoord).then(query => {
-          return query();
-        }).then(rslts => {
-          expect(rslts).to.be.instanceof(Array);
-          rslts.should.have.property('length', 1);
-        });
-      });
-
-      it("findIntersects() should find entity matching GeometryCollection coordinates", () => {
-        return geo.findIntersects(geoJSONGeometry).then(query => {
-          return query();
-        }).then(rslts => {
-          expect(rslts).to.be.instanceof(Array);
-          rslts.should.have.property('length', 1);
-        });
-      });
-    });
-
-    describe("Model Geo class findWithin() method with .where() clause", () => {
-      var geo, obj;
-
-      const record = {
-        'name': "whatever",
-        'location': geoJSON };
-
-      beforeEach(() => {
-        geo = storage.getModel('geo');
-        return geo.create(record).then(rslt => {
-          obj = rslt;
-        });
-      });
-      afterEach(() => {
-        return obj.destroy();
-      });
-
-      it("findWithin() should find entity within coordinates", () => {
-        return geo.findWithin(surroundCoord).then(query => {
-          return query().where({ name: "whatever" });
-        }).then(rslts => {
-          expect(rslts).to.be.instanceof(Array);
-          rslts.should.have.property('length', 1);
-        });
-      });
-
-      it("findWithin() should not find entity intersecting coordinates", () => {
-        return geo.findWithin(intersectCoord).then(query => {
-          return query().where({ name: "whatever" });
-        }).then(rslts => {
-          expect(rslts).to.be.instanceof(Array);
-          rslts.should.have.property('length', 0);
-        });
-      });
-    });
-
-    describe("Model Geo class getGeometry() and getCentroid() methods", () => {
-      var geo, obj;
-
-      beforeEach(() => {
-        geo = storage.getModel('geo');
-        return geo.create(record).then(rslt => {
-          obj = rslt;
-        });
-      });
-      afterEach(() => {
-        return obj.destroy();
-      });
-
-      it("getGeometry(obj) should find Polygon objects", () => {
-        let geometry = geo.getGeometry(obj);
-        expect(geometry).to.be.instanceof(Object);
-        expect(geometry).to.deep.equal(geoJSONPolygon);
-      });
-
-      it("getGeometry(obj, 'Polygon', 'Point', 'LineString') should find all geometry objects", () => {
-        let geometry = geo.getGeometry(obj, 'Polygon', 'Point', 'LineString');
-        expect(geometry).to.be.instanceof(Object);
-        expect(geometry).to.deep.equal(geoJSONGeometry);
-      });
-
-      it("getCentroid(obj) should find centroid", () => {
-        let centroid = geo.getCentroid(obj);
-        expect(centroid).to.be.instanceof(Object);
-        expect(centroid).to.deep.equal(geoJSONCentroid);
-      });
-    });
-
-    describe("Model Geo class Polygon cleanup", () => {
-      var geo, obj;
-
-      const record = {
-        'location': geoJSONMess };
-
-      beforeEach(() => {
-        geo = storage.getModel('geo');
-        return geo.create(record).then(rslt => {
-          obj = rslt;
-        });
-      });
-      afterEach(() => {
-        return obj.destroy();
-      });
-
-      it("getGeometry(obj, 'Polygon', 'Point', 'LineString') should return cleaned geometry objects", () => {
-        let geometry = geo.getGeometry(obj, 'Polygon', 'Point', 'LineString');
-        expect(geometry).to.be.instanceof(Object);
-        expect(geometry).to.deep.equal(geoJSONCleaned);
       });
     });
   });
